V1.0.25.0708B
--Update tới return values funciton.

+++ 30_rollercoaster.py
@@ -1,7 +1,8 @@
 # Using int() to accept Numerical Input
 
 height = input("How tall are you, in inches? ")
-height = int(height)
+# height = '475'
+height = float(height)
 
 if height >= 48:
     print("\nYou're tall enough to ride!")
